Refactor buildPostBody to a shared utility
