+++ app/components/pipeline/modal/confirm-action/util.js
@@ -19,47 +19,3 @@ export function truncateMessage(commitMessage) {
 export function capitalizeFirstLetter(string) {
   return string.charAt(0).toUpperCase() + string.slice(1);
 }
-
-/**
- * Build post body for starting a job
- * @param username
- * @param pipelineId
- * @param job
- * @param event
- * @param parameters
- * @param isFrozen
- * @param reason
- * @returns {{causeMessage: string, pipelineId}}
- */
-export function buildPostBody(
-  username,
-  pipelineId,
-  job,
-  event,
-  parameters,
-  isFrozen,
-  reason
-) {
-  const data = {
-    pipelineId,
-    causeMessage: `Manually started by ${username}`
-  };
-
-  if (job.status) {
-    data.startFrom = job.name;
-    data.groupEventId = event.groupEventId;
-    data.parentEventId = event.id;
-  } else {
-    data.startFrom = '~commit';
-  }
-
-  if (parameters) {
-    data.meta = { parameters };
-  }
-
-  if (isFrozen) {
-    data.causeMessage = `[force start]${reason}`;
-  }
-
-  return data;
-}
